+++ src/types/async-iterable.d.ts
@@ -7,10 +7,12 @@ export type AsyncSourceIterable<T> = null | undefined | AsyncDefinedSourceIterab
 type _AsyncIterable<T> = AsyncIterable<T>;
 export { _AsyncIterable as AsyncIterable };
 
+export type AsyncIteratorResult<T> = Promise<IteratorResult<T>>;
+
 export interface AsyncResultIterable<T> {
-  next(value?: any): Promise<IteratorResult<T>>;
-  return(value?: any): Promise<IteratorResult<T>>;
-  throw(e?: any): Promise<IteratorResult<T>>;
+  next(value?: any): AsyncIteratorResult<T>;
+  return(value?: any): AsyncIteratorResult<T>;
+  throw(e?: any): AsyncIteratorResult<T>;
   [Symbol.asyncIterator](): this;
 }